Add opencv 4.11.0 and patch (#5)
Add opencv 4.11.0
Migrate opencv 3.2 patch to opencv 4.11

The patch is a required adaptation of 0001-export-exif-orientation.patch for opencv 4.11
The patch changes:

ImageEncoder and ImageDecoder are now not a pointer to a base class but a class wrapping an instance of the encoder/decoder. This allows a more granular control on when to perform reads, which is a key part of lilliput. With this, we can perform only a readHeader without having to go through all the data, e.g. readData, as opencv would do with imencode and imdecode
Uses a Mat for buffer instead of a ArrayInput/Output to support lilliput Mat handling
new() will return an instance of ImageEncoder/Decoder and not a pointer to the base class anymore
Adds exif orientation when reading headers for jpeg, png and avif